home work 2 version 1.0.0 Async console output
