+++ app.js
@@ -1,14 +1,16 @@
-var express = require('express');
-var app = express();
+// task #1
+
+const express = require('express');
+const app = express();
 
 
 app.get('/', function (req, res) {
     const moment = require('moment');
-    const timerId = setInterval(() => console.log(moment().utc()), 1000);
+    const timerId = setInterval(() => console.log(moment().utc()), process.env.INTERVAL);
     setTimeout(()  => {
         clearInterval(timerId);
         res.send(moment().utc());
-    }, 10000);
+    }, process.env.TIMEOUT);
 });
 
 app.listen(3000, function () {
