+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "nodemon app.js",
+    "start": "INTERVAL=1000 TIMEOUT=10000 node --experimental-modules app.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
